Move JsonTable and tests into their own files under components
